Send out a registration state change signal when the state becomes Searching.
Back in commit 4363907e51d1f956229792c9a987a19825bea299, a change was
made to have the DataCapabilities handler check the registration status,
and to send out a state change signal if the state became Idle. This
was based on the observation that loss of registration doesn't always
result in a RegistrationStateChange callback as one would expect, but
that loss of registration does result in a DataCapabilities callback.
That change was incomplete, because a loss of registration will usually
result in a transition to the Searching state, rather than going
directly to idle.

For good measure, we also check for Denied and Unknown registration
status, although I don't expect we'd see those in this context.

BUG=none
TEST=This is most easily tested using a SIM for which PIN locking is
enabled. Start out with the SIM unlocked, and with the modem registered
on a network. Then, try to change the PIN, or to disable PIN locking,
but give an incorrect PIN. After the third incorrect PIN entry, the
SIM will be blocked and will require the PUK to unblock it. At that point,
the device will become unregistered from the network (and enter the
Searching state). The flimflam service representing the network should
disappear. Note that because of a separate bug in Chrome, the UI will
not update to reflect the fact that the service has disappeared, but
list-services on the command line will show that it's gone. When you
enter the correct PUK, the device should re-register, and the service
should reappear.

Change-Id: Id4bc3b074c98dfb3b1c366cd1448a0e1cb1883d2
Reviewed-on: https://gerrit-int.chromium.org/3275
Tested-by: Eric Shienbrood <ers@google.com>
Reviewed-by: Eric Shienbrood <ers@google.com>