Call blockdev after install_gpt to protect against partition changes.
The partition scheme of the SSD may be detected while the recovery kernel
boots. When chromeos-install writes a new image to the SSD and overwrites
the old partition table, the kernel may not automatically pick up the fact
that the partitions aren't in the same place. When it tries to mount the
stateful partition to put the .developer_mode file there, it will likely not
find a valid filesystem in that case and will fail.

BUG=chromium-os:17489
TEST=Built a base image and a recovery image which exposed the issue. The
.dev_mode file was deleted from /root on the recovery image so that the file
in the stateful partition would be checked for. Then repeated the experiment
with this change and saw that the .developer_mode file was put in the right
place and that the stateful partition was left alone like it was supposed to
be.

Change-Id: I208cb2beb37e1613e0fffe51adc49a05913435ad
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4283
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>