crosh ssh: Clear params if there aren't any.
If there's a space in a line input to ssh, it splits it into cmd and params.
If not, it sets cmd, but leaves params set to whatever it used to be.
This makes "port 2222" followed by "key" with no filename give the odd error
message that there is no key named 2222.

This patch forces params to the empty string if the line does not contain a
space.

BUG=None
TEST=Adhoc (see above test case)

Change-Id: I99e958bbcbbd8ad138ce8dd919019207035f4466
Reviewed-on: http://gerrit.chromium.org/gerrit/4306
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>