cron.conf, cron-lite.conf: remove cron, add cron-lite
Cron is disabled by this change as it is only used for log rotation.

A new upstart config is introduced that handles a bare bones cron
experience which avoids any additional complexity added by the existing
cron daemon.  If additional functionality is needed down the line, we
should revisit this.

TEST=x86-alex logs are cleaned
BUG=chromium-os:17817

Change-Id: Id7b60fae54059adcf6bae1126e0c756868dbcaa6
Reviewed-on: http://gerrit.chromium.org/gerrit/4315
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>