Handle disable while connected the same way modem-manager does.
Changed the disable operation to first do a disconnect, matching the
behavior of modem-manager. Previously, Disable would cancel an
in-progress connection attempt, but it wouldn't stop an existing
data session. As a result of this change, the gobi cromo state
machine comes closer to matching modem-manager's state machine.

With this change, it should be possible to remove the code from
flimflam that does a disconnect before carrying out a disable
request.

BUG=chromium-os:12572
TEST=While connected, ran mm-disable in a shell, and observed that
the Power property shown in list-devices changed to 'false', and
that in the UI on the settings page, the cellular network disappeared
and the "Disable Mobile Data" button changed to "Enable Mobile Data".

Change-Id: Id16179827dd5249d02a7c19293082a2e0d835730
Reviewed-on: https://gerrit-int.chromium.org/3364
Reviewed-by: Jason Glasgow <jglasgow@google.com>
Tested-by: Eric Shienbrood <ers@google.com>