Handle ForceDisconnect failures gracefully.
Handle ForceDisconnect failures gracefully by immediately disabling
the modem instead of waiting for what may be 5 minutes for the
StartDataSession command to finally timeout.

BUG=chrome-os-partner:4823

TEST=Ensure that Autoconnect is True.  Suspend machine. Resume
machine, and quickly suspend.  Timing matters.  The ideal amount of
time to wait before suspending for the second time is 5 - 15 seconds.
Certainly you want to suspend before the 3G connection is established,
but you need to wait long enough for the modem firmware to download
and for flimflam to start the connect.

Change-Id: I26f6e92f11c917de2fdd46193db2d6618b9fcd5a
Reviewed-on: https://gerrit-int.chromium.org/3373
Reviewed-by: Jason Glasgow <jglasgow@google.com>
Tested-by: Jason Glasgow <jglasgow@google.com>