Provide BIOS information on ARM platforms.
The /usr/sbin/mosys utility is present in the ARM file system,
this fact is considered a reason to use it by the
session_manager_setup.sh. However, on ARM the utility does not do
much, information it is supposed to provide is missing.

This change makes the best of crossystem in case invocation of
`/usr/sbin/mosys -l smbios' resulted in an error.

BUG=chromium-os:15800
TEST=manual

Build updated image for x86-alex and tegra2_kaen. Observe
chrome://system bios_info section on both systems. On x86 there
is no change, on ARM a new line is added "version ". Note that
the "release date", "size" and "ec_version" lines (reported by
mosys on x86) are not being instantiated on the ARM system.

Change-Id: I6648af7088e094a66a621b9b46da2279f6329081
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/4547
Reviewed-by: Chris Masone <cmasone@chromium.org>