Pass the command-line arguments down to /usr/bin/modem as a single argument
Pass the command-line arguments down to /usr/bin/modem as a single
argument, so that "modem_set_carrier Generic UMTS" works like it used
to. This should be considered backwards-compatibility; users of the
functionality in /usr/bin/modem should use quoting as they would for
any other command-line argument with spaces in it.

BUG=chrome-os-partner:5030
TEST=Run "modem_set_carrier Generic UMTS" from crosh or a terminal.

Change-Id: I7038ce36e85f811a3c93399f1e3a075176b60764
Reviewed-on: https://gerrit-int.chromium.org/3458
Tested-by: Nathan J. Williams <njw@google.com>
Reviewed-by: Jason Glasgow <jglasgow@google.com>