dded distinction between DEV_MODE and CROS_DEBUG
Developer mode is represented by the variable DEV_MODE in this script.
CROS_DEBUG is a pseudo-boolean variable that equals to DEV_MODE || DEV_IMAGE.
CROS_DEBUG is 1 if we are in developer mode or we've booted a developer image.
CROS_DEBUG is used to determine if we need to mount /usr/local. DEV_MODE is
used to determine if we need to go through the transition to/from developer
mode and normal mode.

BUG=chromium-os:16648
TEST=Run cros_au_test_harness. Also boot a base image in developer mode and
check for a correct mount of /usr/local. After that, check that the transition
to normal mode goes as expected.

Change-Id: Iee5c89d929f1850d7d4e1a4ff5f21adda800025e
Reviewed-on: http://gerrit.chromium.org/gerrit/4177
Reviewed-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>
Tested-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>