[login_manager] Add the stopping session state
The session_manager used to emit a stopped signal when it began
shutting down a user's session.  Instead, it now emits stopping when
Chrome calls StopSession over DBus, and stopped only after it has shut
down the browser and is ready to exit.

RetrieveState is also updated to behave properly with respect to this
new state.

BUG=chromium-os:17606
TEST=unit

Change-Id: I1d582ab29718537834d4f8b840eaa62ed52e86f6
Reviewed-on: http://gerrit.chromium.org/gerrit/4624
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>