power: Initialize current brightness level at startup.
I was seeing the screen always get dimmed to 10% brightness
whenever powerd got restarted on my Mario device.  The logs
also reported that the initial brightness was 0%, even when
it was actually 100%.  Not sure if the two are related, but
updating the current brightness  member when we start seems
to fix this.

BUG=chromium-os:18206
TEST=manual: set the brightness to various levels, logged in and out, and checked that the brightness level remained constant

Change-Id: Ie39de972beb6b0295457873e21c3055d823845d0
Reviewed-on: http://gerrit.chromium.org/gerrit/4979
Reviewed-by: Simon Que <sque@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>