Refine the ownership and permissions of mount directories.
This CL modifies MountManager to set the ownership and permissions of
mount directories it creates such that only cros-disks and chronos have
read, write and execute permissions on the mount directories while other
users/groups have no permissions at all.

This allows a FUSE-based filesystem daemon (such as ntfs-3g) running as
chronos, instead of root, to have the required permissions on the mount
directories in order to perform mounting.

BUG=chromium-os:18511
TEST=Tested mounting through cros-disks DBus API on Cr48.

Change-Id: Ia41cc8fc4acc4784998b397a6f610e4f5336dc62
Reviewed-on: http://gerrit.chromium.org/gerrit/5131
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>