Allow non-privileged mounts of NTFS filesystem via ntfs-3g.
This CL adds a NTFSMounter class to run ntfs-3g as a non-root user to
allow non-privileged mounts of NTFS filesystem, which is the first step
to limit potential security risks.

It also modifies DiskManager to create a mounter by type identifier,
which provides a better way to add more mounter implementations later.

BUG=chromium-os:18684
TEST=Tested mounting and unmounting a NTFS-formatted USB drive on Cr48.

Change-Id: Ib32177e9273e8569fb225353b3def71257a481c7
Reviewed-on: http://gerrit.chromium.org/gerrit/5303
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>