Post factory wipe-out option(clobber-state)
BUG=chrome-os-partner:4647
TEST=Check if it behaves as we expected after factory wipe-out. In order
to test this, the factory image should write
/mnt/stateful_partition/factory_wipe_option with "battery_cut_off",
"shutdown" or "reboot".

This change will provide factory wipe-out option. The factory image
needs to pass the option through
/mnt/stateful_partition/factory_wipe_option file with one of
"battery_cut_off", "shutdown" or "reboot". If the file doesn't exist it
will fall back to the default option, "reboot".

It contains the typo fix in the patch 1.

Change-Id: I79bb20091f2cc14769ae5828e0b0301a85cb0d0b
Change-Id: I58f8e91b13532dbbfc50a478cc0b8fa46e1df010
Reviewed-on: http://gerrit.chromium.org/gerrit/5198
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Nick Sanders <nsanders@chromium.org>
Tested-by: Jay Kim <yongjaek@chromium.org>