power_manager : Postpone lid action until after powerd respawns
Powerm is currently configured to suspend the system in the absence
of powerd. This may occur due to a bug in powerd, for example.

However, when a user logs out, powerd is stopped until the ui is
restarted typically only a second later. If the user closes the lid
in this brief window after logging out, powerm will assume powerd is
absent and immediately suspend. The expectation is the same
behavior as when we are at the login screen : the system should shut
down.

This change will no longer suspend in this brief window, and instead
allow logout to happen normally and another powerd to start.
Powerm will wait to send the lid close notification to the new powerd.
The end result is that powerd will start, and immediately issue a shutdown.

This change also reorders power daemon's init to before the
power_manager_service init, and removes the dbus_request_name in
suspender for the PowerManager Interface. This allows powerd to respond
to an early LidClosed signal from powerm as soon as the interface is
advertised as being claimed.

BUG=chromium-os:17606
TEST=log in as guest, or a user account.
ctl-shift-q to log out.
Immediately close the lid.

The system should shut down instead of suspend.

Change-Id: I3c0a08999cfd75d31dc77e7d62936cf13dacbae3
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5458
Reviewed-by: Sameer Nanda <snanda@chromium.org>