Add hysteresis to battery remaing time calculation
Do not allow random swings in reported battery discharge time.  The
calculated linear time can vary by a lot.  With this patch, there is a
limit to how much variation is allowed.  Anything outside the variation
range will be clipped.  Only when the time has been outside of the range
for long enough will it be allowed to become the new "baseline" time.

BUG=chromium-os:14506
TEST=Run 'FEATURES=test emerge-x86-mario power_manager' and make sure
the tests pass.

Change-Id: Ib2a99e7d2804fc1d68a7fa9949be96a2e8a7478e
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6565
Reviewed-by: Sameer Nanda <snanda@chromium.org>