power_manager : only check hwclock in powerd_suspend under test
powerd_suspend records the output of /sbin/hwclock at every resume
that isn't canceled. The output is only used by power_Resume,
however, and is causing an issue on ARM platforms where the
read of the hardware rtc will cause a wake event for a subsequent
quick suspend.

This change enables shflags for powerd_suspend, and creates the
flag --test (with notest as default) to enable the read of hwclock.

Also, this formalizes the powerd_suspend cancel case into its own flag,
--cancel.

The appropriate change has also been made in powerman.cc to pass in
--cancel in the cancel case.

BUG=chrome-os-partner:5326
TEST=1.run power_Resume test.
2. Ensure that results include reasonable results based on rtc.
3. tail /var/log/messages -n 50.
4. Ensure "Recording hwclock time for test" is logged in test case.
5. Suspend/Resume the system using the lid.
6. tail /var/log/messages -n 50
7. Ensure "Recording hwclock time for test" is not logged in normal case.

Change-Id: I5b81eedbe58cd1db6342620ff7875bc9b7bb8162
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5711
Reviewed-by: Todd Broch <tbroch@chromium.org>