Move definitions of D-Bus string contants from .cc file to .h file.
This will make it easier to use the constants. All you have to do is just
to include the header file. No need to link libchromeos. The goal is to
share the constants with Chrome.

The patch was created by a Python script, which has been uploaded to crosbug.com/19054

This change will reduce the binary size, as the compiler will be able to
eliminate unused constants, whereas all the constants were embedded into
the binary as part of libchromeos.a, regardless of whether they are used
or not, before this change.

The following is the result from libcros and session_manager (chromeos-login).

-rwxr-xr-x 1 satorux eng 690048 Aug 11 11:22 libcros.so.before
-rwxr-xr-x 1 satorux eng 655788 Aug 11 11:53 libcros.so.after

-rwxr-xr-x 1 satorux eng 865108 Aug 11 16:02 session_manager.before
-rwxr-xr-x 1 satorux eng 855428 Aug 11 16:04 session_manager.after

BUG=chromium-os:19054
TEST=emerge packages that use service_constants.h: libchromeos cromo cros-disks libcros chromeos-cryptohome chromeos-imageburner chromeos-login power_manager shill speech_synthesis chromeos-wm

Change-Id: I83d360a201f8936d76731de255bc8caf2e3ee0be
Reviewed-on: http://gerrit.chromium.org/gerrit/5854
Tested-by: Satoru Takabayashi <satorux@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>