power_manager: monitor_reconfigure: Turn on display when projecting
When reconfiguring monitors on an external monitor attach
event, we blank the screen to hide the resize.

However, we only unblank the screen when there are
resolutions available for the internal display, ignoring
the external display.

This checks also to see if we are presenting and unblanks
the displays in that case as well.

BUG=chromium-os:19146
TEST=Boot chromebook without external monitor attached.
Attach a monitor of higher resolution (1920 x 1200)
via VGA or HDMI.
Ensure that external monitor turns on immediately
on plug.

Change-Id: If203cbf271550b34d9d3b5cbde9d583b2424b523
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5950
Reviewed-by: Simon Que <sque@chromium.org>