Add an option in mount managers to reserve mount paths on errors.
This CL introduces an option in mount managers to reserve a mount path
when the mount operation fails due to certain types of error. This
change is required by the File Browser UI to keep the mount directory
for an unknown or unsupported filesystem on an external storage in order
to allow users to select the directory and format the corresponding
filesystem.

BUG=chromium-os:19469
TEST=Ran unit tests and tested mounting unknown filesystem on Cr48.

Change-Id: I0b680b0730dc344f17284217a30ebb7357b689cc
Reviewed-on: http://gerrit.chromium.org/gerrit/6406
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>