power_manager: monitor_reconfigure: Enable external outputs after resizing the framebuffer
When reconfiguring the screen, monitor_reconfigure first enables the
external display, then resizes the framebuffer.  When the external
display has a higher resolution than the internal display, this causes
the server to temporarily have an invalid configuration where the CRTC
for the external output is scanning out a higher resolution (such as
1920x1080) while the framebuffer still has a smaller resolution (e.g.
1366x768).  This triggers the shadow code in xf86CrtcRotate, which
allocates a 1920x1080 shadow framebuffer and starts copying from the
1366x768 primary framebuffer.  This situation does not go away when
monitor_reconfigure resizes the framebuffer to 1920x1080.

Fix this by resizing the framebuffer to the external display's size
before actually turning the external CRTC on.

BUG=chrome-os-partner:5654
TEST=On Seaboard, hotplug a 1920x1080 HDMI monitor and verify that xf86CrtcRotate does not enable the shadow framebuffer

Change-Id: Ibda811226c3d559f11f84fc033092e29d3b35955
Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/6497
Reviewed-by: Simon Que <sque@chromium.org>