Correct battery time coming out of suspend
When resuming, the reported battery current is very low.  A simple
linear estimate for time remaining results in an extraordinarily high
value.  This patch takes suspend/resume into account when doing the
hysteresis, so that the battery time is normal when coming out of
suspend.

BUG=chromium-os:9606
TEST=Suspend by closing lid. Resume and quickly check the time remaining
by clicking on the battery icon. It should show something similar to
what it showed before suspend.

Change-Id: Ia295ef4a14e59ca581e07a53903e37eb0d81795b
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6874
Reviewed-by: Sameer Nanda <snanda@chromium.org>