Eliminate one cause of "Reentrant SDK access detected" warnings.
Whenever a call to GetSignalStrengthDbm is made, it calls
GetSessionState to determine whether the modem is connected,
and if so, reports the signal strength for the actual data
bearer in use. However, if this happens while a connection
is being attempted, the call to GetSessionState fails because
it is in the same SDK "service" as StartDataSession. The
solution is to instead check the session_id_ member variable
to determine whether the modem is connected.

BUG=chromium-os:14100
TEST=In the UI, press the connect button, and in a terminal
window, immediately run "mm-status" (or "modem status" in crosh).
With the fix, there should *not* be a log message that says "Reentrant
SDK access detected: Called GetSessionState while already in call to
StartDataSession".

Change-Id: I4f73d7128c8afa702dc55945aecc88a09bb9fb1c
Reviewed-on: https://gerrit-int.chromium.org/4799
Tested-by: Eric Shienbrood <ers@google.com>
Reviewed-by: Eric Shienbrood <ers@google.com>