Cleanup comments, dead test code, new marker file.
This code is a workaround for a bug in our startup ordering between a
writable data area being available to check for a marker file vs.
starting up the udevd.  The bug manifests as the udevd winning the race
vs. the mount and creating a serio_raw0 device and putting the mouse
i8042 PS/2 interface into raw mode when the .tp_disable file is present.

Resolving this bug correctly would require parameter NVRAM, which would
potentially result in a security issue, or changing the startup order to
ensure mounting before udevd runs, which would result in a startup mount
regression.

BUG=chromium-os:20190
TEST=Run code on ZGB

Change-Id: I0e786806bf2a5b67d8253924d4b0a639bcfcff66
Reviewed-on: http://gerrit.chromium.org/gerrit/7413
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Terry Lambert <tlambert@chromium.org>