Change the handling of parent block devices.
This CL modifies the handling of parent block devices such that what
used to be considered non-automountable are now reported as
automountable but marked as hidden so that the file browser does not
show and mount them, but the image burner can see and use them.

BUG=chromium-os:20476
TEST=Verified parent block devices are reported by
EnumerateAutoMountableDevices and are marked as hidden by
GetDeviceProperties if they contain at least one partition.

Change-Id: Ia005a3ff1b3fee20add034c7cbac5133268499de
Reviewed-on: http://gerrit.chromium.org/gerrit/7941
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>