power_supply: after suspend, adjust last poll time
BUG=chromium-os:20520
TEST=To monitor the time to empty, type:
'tail -f /var/log/power_manager/powerd.LATEST | grep TimeToEmpty'
Note the time that is being reported.  Let the system run for a few
minutes for the time to stabilize.  It is better to do this on AC power.
The suspend for the amount of
time that is reported, plus an hour.  e.g. if time=300 min, suspend for
360 minutes.  Or just leave it overnight.  Then resume.  The TimeToEmpty
should not have dropped off significantly.  If it goes negative, it has
failed.

Change-Id: Icc59ad785a5e9ce8350de6daa8da3b60c336b967
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8056
Reviewed-by: Sameer Nanda <snanda@chromium.org>