Unmount all removable devices or reserved mount points when a session ends.
DiskManager currently reserves a virtual mount point for a device with
an unrecognized or unsupported filesystem and notifies Chrome an error.
When the user logs out and logs back in again, the virtual mount point
is preserved across sessions, but the error information is not. Chrome
thus cannot differentiate between a real mount point and a virtual one,
and mistakenly treat the virtual mount point as real. This CL resolves
this issue by forcing DiskManager to unmount all its managed mount
points when a session ends.

BUG=chromium-os:21111
TEST=Tested the following on a Cr48:
1. Log in to a new session and then plug in a USB drive.
2. Verify that the USB drive is mounted to a folder under External
   Storage in the file browser.
3. Log out the session.
4. ssh into the Chromebook and verify that no directory remains in
   /media/removable.
5. Log in to a new session again.
6. Verify that the USB drive is again mounted to a folder under
   External Storage in the file browser.

Change-Id: I49185c8129530d500ff598859858849406c08b58
Reviewed-on: http://gerrit.chromium.org/gerrit/8581
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>