Change log pattern for detecting IPsec authentication failures.
The previously used log pattern for detecting IPsec authentication
failures did not work when connnecting to a Windows 2008 RRAS server or
Cisco ASA 5505. This CL changes the log pattern to better handle these
cases.

BUG=chromium-os:18573
TEST=Tested connecting to Windows 2008 RRAS server, Cisco ASA 5505 and
StrongSWAN VPN server with invalid pre-shared key and certificate, and
verified that the UI shows the expected error messages.

Change-Id: I20fbaeecaea59f556321d2e799e1de213a8eb947
Reviewed-on: http://gerrit.chromium.org/gerrit/8609
Reviewed-by: Ken Mixter <kmixter@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>