power: Work around race when locking screen.
When the power button is being held, there's a delay between
when powerd asks the session manager to lock the screen and
when it receives notification from Chrome that the screen is
locked.  When the power button was pressed again during this
interval, powerd would again ask the session manager to lock
the screen and tell the window manager to display the
pre-lock animation, which would result in some visible jank.

This change makes powerd avoid sending a second lock request
and instead wait for notification that the screen has been
locked before beginning the pre-shutdown sequence.

BUG=chromium-os:21137
TEST=manual: patched the session manager and libcros to delay sending back notification when the screen is locked

Change-Id: I6590162d89868fb126165aa852283fc399dd8d1d
Reviewed-on: http://gerrit.chromium.org/gerrit/8677
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>