cromo: don't fail post-stop if modem disappears
cromo's post-stop script is designed to restart modems that are left in
a potentially bad state when cromo terminates abnormally.  However, it's
possible for the modem to disappear between cromo terminating and the
script trying to reboot it.  If the script fails, upstart will assume it
is unsafe to restart cromo, and 3G won't work until the machine reboots.

This change checks whether the device exists in sysfs before attempting
to restart it, and further ensures that the script will succeed even if
the echo 0 or echo 1 statements fail (in the extremely rare case that
the modem disappears between the earlier check and the echo, or between
the echoes).  It also logs a little more, to help later debugging.

BUG=chromium-os:21463
TEST=cromo comes back up when performing steps in bug report

Change-Id: Ib98fcaa72a2b9340d6227bea73b44755ba2aefee
Reviewed-on: http://gerrit.chromium.org/gerrit/9860
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Commit-Ready: ttuttle <ttuttle@chromium.org>
Reviewed-by: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>